# docs/swagger.yaml
definitions:
  model.CreateSubscriptionRequest:
    properties:
      end_date:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        maxLength: 255
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - price
    - service_name
    - start_date
    - user_id
    type: object
  model.SubscriptionSwagger:
    description: Subscription data
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  model.SummaResponse:
    properties:
      total_price:
        type: integer
    type: object
  model.UpdateSubscriptionRequest:
    properties:
      end_date:
        type: string
      id:
        type: string
      price:
        minimum: 0
        type: integer
      service_name:
        maxLength: 255
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - id
    type: object
info:
  contact: {}
paths:
  /subscriptions:
    get:
      consumes:
      - application/json
      description: Returns list of subscriptions with optional filtering by service_name,
        user_id, start_date, end_date
      parameters:
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      - description: Filter by user UUID
        in: query
        name: user_id
        type: string
      - description: Filter by start date (MM-YYYY)
        in: query
        name: start_date
        type: string
      - description: Filter by end date (MM-YYYY)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of subscriptions
          schema:
            items:
              $ref: '#/definitions/model.SubscriptionSwagger'
            type: array
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Creates a new subscription. Checks for duplicate service_name,
        user_id, start_date combination.
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Subscription created
          schema:
            $ref: '#/definitions/model.SubscriptionSwagger'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: Subscription already exists
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      consumes:
      - application/json
      description: Deletes a subscription by its UUID
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Invalid UUID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Subscription not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      consumes:
      - application/json
      description: Returns a single subscription by its UUID
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Subscription found
          schema:
            $ref: '#/definitions/model.SubscriptionSwagger'
        "400":
          description: Invalid UUID
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Subscription not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Updates an existing subscription by its UUID
      parameters:
      - description: Subscription UUID
        in: path
        name: id
        required: true
        type: string
      - description: Subscription update data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Subscription updated
          schema:
            $ref: '#/definitions/model.SubscriptionSwagger'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Subscription not found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/summa:
    get:
      consumes:
      - application/json
      description: Returns the sum of all price values for subscriptions matching
        the filters
      parameters:
      - description: Filter by service name
        in: query
        name: service_name
        type: string
      - description: Filter by user UUID
        in: query
        name: user_id
        type: string
      - description: Filter by start date (MM-YYYY)
        in: query
        name: start_date
        type: string
      - description: Filter by end date (MM-YYYY)
        in: query
        name: end_date
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Total price
          schema:
            $ref: '#/definitions/model.SummaResponse'
        "400":
          description: Invalid request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get summa of subscription prices
      tags:
      - subscriptions
swagger: "2.0"